+++ mypy/expandtype.py
@@ -1,8 +1,9 @@
 from __future__ import annotations
 
-from typing import Final, Iterable, Mapping, Sequence, TypeVar, cast, overload
+from contextlib import contextmanager
+from typing import Final, Generator, Iterable, Mapping, Sequence, TypeVar, cast, overload
 
-from mypy.nodes import ARG_STAR, FakeInfo, Var
+from mypy.nodes import ARG_STAR, CONTRAVARIANT, COVARIANT, FakeInfo, Var, INVARIANT
 from mypy.state import state
 from mypy.types import (
     ANY_STRATEGY,
@@ -38,6 +39,7 @@ from mypy.types import (
     UninhabitedType,
     UnionType,
     UnpackType,
+    VarianceModifier,
     flatten_nested_unions,
     get_proper_type,
     split_with_prefix_and_suffix,
@@ -53,37 +55,49 @@ import mypy.type_visitor  # ruff: isort: skip
 
 
 @overload
-def expand_type(typ: CallableType, env: Mapping[TypeVarId, Type]) -> CallableType: ...
+def expand_type(
+    typ: CallableType, env: Mapping[TypeVarId, Type], *, variance: int | None = ...
+) -> CallableType: ...
 
 
 @overload
-def expand_type(typ: ProperType, env: Mapping[TypeVarId, Type]) -> ProperType: ...
+def expand_type(
+    typ: ProperType, env: Mapping[TypeVarId, Type], *, variance: int | None = ...
+) -> ProperType: ...
 
 
 @overload
-def expand_type(typ: Type, env: Mapping[TypeVarId, Type]) -> Type: ...
+def expand_type(
+    typ: Type, env: Mapping[TypeVarId, Type], *, variance: int | None = ...
+) -> Type: ...
 
 
-def expand_type(typ: Type, env: Mapping[TypeVarId, Type]) -> Type:
+def expand_type(typ: Type, env: Mapping[TypeVarId, Type], *, variance=None) -> Type:
     """Substitute any type variable references in a type given by a type
     environment.
     """
-    return typ.accept(ExpandTypeVisitor(env))
+    return typ.accept(ExpandTypeVisitor(env, variance=variance))
 
 
 @overload
-def expand_type_by_instance(typ: CallableType, instance: Instance) -> CallableType: ...
+def expand_type_by_instance(
+    typ: CallableType, instance: Instance, *, use_variance: int | None = ...
+) -> CallableType: ...
 
 
 @overload
-def expand_type_by_instance(typ: ProperType, instance: Instance) -> ProperType: ...
+def expand_type_by_instance(
+    typ: ProperType, instance: Instance, *, use_variance: int | None = ...
+) -> ProperType: ...
 
 
 @overload
-def expand_type_by_instance(typ: Type, instance: Instance) -> Type: ...
+def expand_type_by_instance(
+    typ: Type, instance: Instance, *, use_variance: int | None = ...
+) -> Type: ...
 
 
-def expand_type_by_instance(typ: Type, instance: Instance) -> Type:
+def expand_type_by_instance(typ: Type, instance: Instance, use_variance=None) -> Type:
     """Substitute type variables in type using values from an Instance.
     Type variables are considered to be bound by the class declaration."""
     if not instance.args and not instance.type.has_type_var_tuple_type:
@@ -108,12 +122,11 @@ def expand_type_by_instance(typ: Type, instance: Instance) -> Type:
         else:
             tvars = tuple(instance.type.defn.type_vars)
             instance_args = instance.args
-
         for binder, arg in zip(tvars, instance_args):
             assert isinstance(binder, TypeVarLikeType)
             variables[binder.id] = arg
 
-        return expand_type(typ, variables)
+        return expand_type(typ, variables, variance=use_variance)
 
 
 F = TypeVar("F", bound=FunctionLike)
@@ -181,10 +194,28 @@ class ExpandTypeVisitor(TrivialSyntheticTypeTranslator):
 
     variables: Mapping[TypeVarId, Type]  # TypeVar id -> TypeVar value
 
-    def __init__(self, variables: Mapping[TypeVarId, Type]) -> None:
+    def __init__(
+        self, variables: Mapping[TypeVarId, Type], *, variance: int | None = None
+    ) -> None:
         super().__init__()
         self.variables = variables
         self.recursive_tvar_guard: dict[TypeVarId, Type | None] = {}
+        self.variance = variance
+        self.using_variance: int | None = None
+
+    @contextmanager
+    def in_variance(self) -> Generator[None]:
+        using_variance = self.using_variance
+        self.using_variance = CONTRAVARIANT
+        yield
+        self.using_variance = using_variance
+
+    @contextmanager
+    def out_variance(self) -> Generator[None]:
+        using_variance = self.using_variance
+        self.using_variance = COVARIANT
+        yield
+        self.using_variance = using_variance
 
     def visit_unbound_type(self, t: UnboundType) -> Type:
         return t
@@ -238,6 +269,19 @@ class ExpandTypeVisitor(TrivialSyntheticTypeTranslator):
         if t.id.is_self():
             t = t.copy_modified(upper_bound=t.upper_bound.accept(self))
         repl = self.variables.get(t.id, t)
+        use_site_variance = repl.variance if isinstance(repl, VarianceModifier) else None
+        positional_variance = self.using_variance or self.variance
+        if (
+            positional_variance is not None
+            and use_site_variance is not None
+            and use_site_variance is not INVARIANT
+            and positional_variance != use_site_variance
+        ):
+            repl = (
+                t.upper_bound.accept(self)
+                if positional_variance == COVARIANT
+                else UninhabitedType()
+            )
         if isinstance(repl, ProperType) and isinstance(repl, Instance):
             # TODO: do we really need to do this?
             # If I try to remove this special-casing ~40 tests fail on reveal_type().
@@ -414,10 +458,15 @@ class ExpandTypeVisitor(TrivialSyntheticTypeTranslator):
             needs_normalization = True
             arg_types = self.interpolate_args_for_unpack(t, var_arg.typ)
         else:
-            arg_types = self.expand_types(t.arg_types)
+            with self.in_variance():
+                arg_types = self.expand_types(t.arg_types)
+        with self.out_variance():
+            ret_type = t.ret_type.accept(self)
+            if isinstance(ret_type, VarianceModifier):
+                ret_type = ret_type.value
         expanded = t.copy_modified(
             arg_types=arg_types,
-            ret_type=t.ret_type.accept(self),
+            ret_type=ret_type,
             type_guard=t.type_guard and cast(TypeGuardType, t.type_guard.accept(self)),
             type_is=(t.type_is.accept(self) if t.type_is is not None else None),
         )
@@ -538,7 +587,10 @@ class ExpandTypeVisitor(TrivialSyntheticTypeTranslator):
     def expand_types(self, types: Iterable[Type]) -> list[Type]:
         a: list[Type] = []
         for t in types:
-            a.append(t.accept(self))
+            typ = t.accept(self)
+            if isinstance(typ, VarianceModifier):
+                typ = typ.value
+            a.append(typ)
         return a